Towers and main bullet
Added a tower that buffs the bullets damage when the bullet collides with it.

Changed the behavior of the duplicator tower. It now spawns a bullet in the oposite direction the bullet collided into.

The bullets always explode now.

The bullets damage decays over time.

The bullet gets smaller as its damage decays.